Adds My Suggestions to mass consensus flow
Enables admins to include a "My Suggestions" step in the mass consensus flow, allowing users to view and manage their own suggestions during the process.

This integration streamlines the user experience and provides admin control over the inclusion and placement of the "My Suggestions" step within the flow. Also includes updating delib-npm to the newest version.