Refactors participant count and statement fetching
Improves performance of results summary by fetching the total number of participants directly from the parent statement's evaluation data.

This removes the need to iterate through sub-statements to calculate the participant count and listen to all substatements, and instead listens only to the main statement and leverages the `asParentTotalEvaluators` field in the evaluation data to display participant count.