Adds test-minified build and deploy targets
Introduces a new build and deploy target specifically for minified testing builds.

This allows for testing the production-like, minified version of the application in a testing environment before deploying to production.

Updates the build process to allow for minification or not based on the target environment. Sourcemaps are only generated for non-minified test builds to improve debugging.